[examples] Fix, sf population polygons div by zero error (#8209)

+++ superset/examples/sf_population_polygons.py
@@ -17,7 +17,7 @@
 import json
 
 import pandas as pd
-from sqlalchemy import BigInteger, Text
+from sqlalchemy import BigInteger, Float, Text
 
 from superset import db
 from superset.utils import core as utils
@@ -43,7 +43,7 @@ def load_sf_population_polygons(only_metadata=False, force=False):
                 "zipcode": BigInteger,
                 "population": BigInteger,
                 "contour": Text,
-                "area": BigInteger,
+                "area": Float,
             },
             index=False,
         )
